SSAP/requirements.txt
asgiref==3.8.1
autopep8==2.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
dj-rest-auth==6.0.0
Django==4.2
django-allauth==0.63.1
django-extensions==3.2.3
django-seed==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==25.2.0
idna==3.7
pillow==10.3.0
pycodestyle==2.11.1
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.32.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
toposort==1.10
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
